server/requirements.txt
django==2.2.7
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-sso==0.3.2
django-postgrespool==0.3.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pikepdf==1.15.0
pyAesCrypt==0.4.3
django-mptt==0.11.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-import-export==2.0.
django-jet==1.0.8
requests==2.23.0
python-dateutil==2.8.1